Breakout an alignments and bigwig test file
